[jw5022][aw1122] - Fixed compile errors due to float usage for recent_cpu member, fixed MULT_FP_BY_INT call

+++ src/threads/thread.h
@@ -99,7 +99,7 @@ struct thread
     struct list_elem elem;              /* List element. */
 
     int nice;                           /* Niceness value. */
-    float recent_cpu;                   /* recent_cpu value. */
+    int64_t recent_cpu;                   /* recent_cpu value. */
 
 #ifdef USERPROG
     /* Owned by userprog/process.c. */
